New functions added

+++ selectsoft_app/app/Http/Controllers/UserController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\RegisterRequest;
-use App\Http\Requests\UserUpdataRequest;
+use App\Http\Requests\UpdatePasswordRequest;
 use App\Mail\WelcomeMailable;
 use App\Models\Candidate;
 use App\Models\City;
@@ -15,6 +15,7 @@ use App\Models\Recruiter;
 use App\Models\Role;
 use App\Models\Selector;
 use App\Models\User;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
@@ -70,21 +71,19 @@ class UserController extends Controller
         $newUser->password = $request->input('password');
         $newUser->role_id = 1;
 
-        $newUser->save();
 
+        $mail = $request->email;
 
-        $mail = $newUser->email;
-
-        $userName = $newUser->name." ".$newUser->last_name;
+        $userName = $request->name." ".$request->last_name;
 
         Mail::to($mail)->send(new WelcomeMailable($userName));
 
+        $newUser->save();
 
         if($newUser->role_id == 1) {
             $newCandidate = new Candidate();
             $newCandidate->user_id = $newUser->id;
             $newCandidate->occupational_profile = 'NULL';
-
 
             $newCandidate->save();
         }
@@ -93,6 +92,28 @@ class UserController extends Controller
 
     }
 
+    public function updatePassword() {
+        $user = Auth::user();
+        $role_id = $user->role_id;
+
+        return view('/update_password/updatePassword', [
+            'user' => $user,
+            'role_id' => $role_id
+        ]);
+    }
+
+    public function storePassword(UpdatePasswordRequest $request) {
+        $user = Auth::user();
+        $userToUpdate = User::find($user->id);
+
+        $userToUpdate->password = $request->password;
+        $userToUpdate->save();
+
+        auth()->logout();
+
+        return redirect()->route('user.login');
+
+    }
 
     public function editUserRole(User $user) {
         $userAuth = Auth::user();
@@ -114,17 +135,26 @@ class UserController extends Controller
         ]);
     }
 
-    public function updateUserRole(User $userMod, Request $request) {
+    public function updateUserRole(User $userMod, Request $request) :RedirectResponse {
         if($userMod->role_id ==1){
             $candidate = Candidate::where('user_id', $userMod->id)->first();
 
             $candidate->delete();
         }else if ($userMod->role_id == 2) {
             $selector = Selector::where('user_id', $userMod->id)->first();
-            $selector->delete();
+            if ($selector->company_id == null){
+                $selector->delete();
+            }else{
+                return redirect()->route('instructor.selectors')->with('message', 'El seleccionador aun esta asociado a una empresa');
+            }
         } else if($userMod->role_id == 3){
             $recruiter = Recruiter::where('user_id', $userMod->id)->first();
-            $recruiter->delete();
+            if($recruiter->company_id == null){
+                $recruiter->delete();
+            }else{
+                return redirect()->route('instructor.recruiters')->with('message', 'El reclutador aun esta asociado a una empresa');
+            }
+
         } else if($userMod->role_id == 4){
             $instructor = Instructor::where('user_id', $userMod->id)->first();
             $instructor->delete();
@@ -171,68 +201,18 @@ class UserController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit()
+    public function edit(string $id)
     {
-        $user = Auth::user();
-        $role_id = $user->role_id;
-        $countries = Country::all();
-        $cities = City::all();
-        $departaments = Departament::all();
-        $document_types = Document_type::all();
-        $roles = Role::all();
-        return view('/update_data/update_data',[
-            'user' => $user,
-            'role_id' => $role_id,
-            'countries' => $countries,
-            'document_types' => $document_types,
-            'cities' => $cities,
-            'departaments' => $departaments,
-            'roles' => $roles
-        ]);
-
+        //
     }
-
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UserUpdataRequest $request)
+    public function update(Request $request, string $id)
     {
-        $user=Auth::user();
-        $newUser=User::find($user->id);
-        $newUser->name = $request->input('name');
-        $newUser->last_name = $request->input('last_name');
-        $newUser->document_type_id = $request->input('document_type_id');
-        $newUser->telephone = $request->input('telephone');
-        $newUser->phone_number = $request->input('phone_number');
-        $newUser->address = $request->input('address');
-        $newUser->id_country = $request->input('id_country');
-        $newUser->id_department = $request->input('id_department');
-        $newUser->id_city = $request->input('id_city');
-
-        $newUser->save();
-        return redirect()->route('user.index');
+        //
     }
-
-    public function newPassword(){
-        $user = Auth::user();
-        $role_id = $user->role_id;
-        return view('/update_password/update_password',[
-            'user' => $user,
-            'role_id' => $role_id
-        ]);
-    }
-
-    public function updatePassword(Request $request){
-
-        $user=Auth::user();
-        $usertoupdate=User::find($user->id);
-        $usertoupdate->password = $request->password;
-        $usertoupdate->save();
-        auth()->logout();
-        return redirect()->route('user.login');
-    }
-
 
     /**
      * Remove the specified resource from storage.
